refactored session links
